Plan 044 Step 1: extract export/CSV conversion to export-format.ts
Move convertToExportFormat, convertToCSV, and getMimeType out of
drizzle-adapter.ts so export logic is a standalone module with CSV
formula-injection escaping preserved unchanged.
